% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempetell.R
\name{gps}
\alias{gps}
\alias{location}
\title{This is the location function alias gps.
It utilizes a google API with key provided .
It gets the location of the user in terms of latitude and longitude.}
\usage{
location()
}
\value{
A list of latitude and of longitude
}
\description{
This is the location function alias gps.
It utilizes a google API with key provided .
It gets the location of the user in terms of latitude and longitude.
}
