% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempetell.R
\name{stations}
\alias{stations}
\alias{PopulateList}
\title{This is the populate list function alias stations.
It utilizes the coordinates generated by the location function.
It gets data from the weather API and generate a dataframe with weather data}
\usage{
PopulateList(gps)
}
\arguments{
\item{gps}{(these are the coordinates from the location function)}
}
\value{
A data-frame with a minimum of twenty nearest weather stations according to user location .
Weather data from the weather stations generated.
}
\description{
This is the populate list function alias stations.
It utilizes the coordinates generated by the location function.
It gets data from the weather API and generate a dataframe with weather data
}
