% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempetell.R
\name{temptell}
\alias{temptell}
\alias{stationMap}
\title{This is the station maps function alias temptell.
It utilizes the returns from the stations and gps functions.
It creates a leaflet map with the weather stations and temperature data.}
\usage{
stationMap(stations, gps)
}
\arguments{
\item{stations}{these are the weather stations retreived from the weather API.}

\item{gps}{These are the latitude and longitude coordinates from the gps function .
It uses the coordinates to set the view of the leaflet map .
It creates the markers for the stations in the map using the same coordinates.}
}
\value{
A leaflet map
Markers for the weather stations with temperature data.
@import leaflet
}
\description{
This is the station maps function alias temptell.
It utilizes the returns from the stations and gps functions.
It creates a leaflet map with the weather stations and temperature data.
}
